\name{mcmc_hkevp}
\alias{mcmc_latent}
\alias{mcmc_hkevp}
\alias{mcmc_deponly}
\title{
C routine for fitting either the HKEVP or the latent variable model with MCMC.
}
\description{
This set of functions are called by \code{hkevp.fit} or \code{latent.fit} and should not be used directly by the user (high risk of segmentation fault)!
}
\author{Quentin Sebille}
\seealso{\code{\link{hkevp.fit}}}
